+++ LEDGeometry/WaveEffect.cpp
@@ -13,13 +13,13 @@ WaveEffect::WaveEffect(CRGB* spectrum, uint8_t resolution, uint8_t* projection,
       resolution(resolution),
       projection(projection),
       frequency(frequency) {
-    for (uint8_t i = 0; i < resolution; i++) {
+    for (int i = 0; i < resolution; i++) {
         spectrum[i] = CRGB::Black;
     }
 }
 
 void WaveEffect::update(LEDCurve* led_curve) {
-    for (uint8_t i = 0; i < led_curve->shape->n_points(); i++) {
+    for (int i = 0; i < led_curve->shape->n_points(); i++) {
         led_curve->leds[i] = spectrum[projection[i]];
     }
     propagate();
@@ -30,7 +30,7 @@ void WaveEffect::update(LEDCurve* led_curve) {
 }
 
 void WaveEffect::propagate() {
-    for (uint8_t i = resolution - 1; i > 0; i--) {
+    for (int i = resolution - 1; i > 0; i--) {
         spectrum[i] = spectrum[i - 1];
     }
 }
